+++ server/mongodb/models/GroupThread.js
@@ -32,9 +32,12 @@ const GroupThreadSchema = new mongoose.Schema({
 async function handleDelete(provDoc) {
   const doc =
     this.getQuery != null ? await this.model.findOne(this.getQuery()) : provDoc;
-  const id = doc._id;
 
-  await mongoose.model("Comment").deleteMany({ parent: id });
+  if (doc != null) {
+    const id = doc._id;
+
+    await mongoose.model("Comment").deleteMany({ parent: id });
+  }
 }
 
 GroupThreadSchema.pre("remove", handleDelete);
